refactor: lay cleanup foundations
Group foundational cleanup for shared types, SDK session state, and orchestrator launcher modules.
