+++ crates/tracepilot-orchestrator/Cargo.toml
@@ -28,6 +28,8 @@ reqwest = { workspace = true }
 
 similar = "2"
 walkdir = "2"
+regex = "1"
+once_cell = "1"
 sha2 = { workspace = true }
 url = "2"
