+++ packages/ui/src/components/SessionCard.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { SessionListItem } from "@tracepilot/types";
-import { formatRelativeTime } from "../utils/formatters";
+import { formatRelativeTime } from "@tracepilot/types";
 import Badge from "./Badge.vue";
 
 defineProps<{
