+++ packages/ui/src/components/ToolDetailPanel.vue
@@ -8,9 +8,9 @@
  * to the parent component.
  */
 import type { TurnToolCall } from "@tracepilot/types";
+import { formatDuration, formatTime } from "@tracepilot/types";
 import { computed } from "vue";
 import { resolveLucideIcon } from "../icons/lucideRegistry";
-import { formatDuration, formatTime } from "../utils/formatters";
 import { extractPrompt, toolIcon } from "../utils/toolCall";
 import Badge from "./Badge.vue";
 import ToolArgsRenderer from "./renderers/ToolArgsRenderer.vue";
