--- packages/ui/src/utils/formatters.ts
@@ -1,27 +1,0 @@
-/**
- * Re-exports from @tracepilot/types for backwards compatibility.
- * All formatting utilities have been moved to @tracepilot/types.
- *
- * This file maintains the same export surface as before, allowing existing
- * code that imports from @tracepilot/ui to continue working unchanged.
- */
-export {
-  formatBytes,
-  formatClockTime,
-  formatCost,
-  formatDate,
-  formatDateMedium,
-  formatDateShort,
-  formatDuration,
-  formatLiveDuration,
-  formatNumber,
-  formatNumberFull,
-  formatPercent,
-  formatRate,
-  formatRelativeTime,
-  formatShortDate,
-  formatTime,
-  formatTokens,
-  toErrorMessage,
-  truncateText,
-} from "@tracepilot/types";
